SystemPage.xaml Version Alineacion de botones en general TechSupportPage.xaml alineaciones de logo y textos, link para correo. Actualizacion de logos de wibsar y retail
