+++ View/CategoryListPage.xaml
@@ -70,7 +70,7 @@
                         <ColumnDefinition Width="*"></ColumnDefinition>
                         <ColumnDefinition Width="*"></ColumnDefinition>
                     </Grid.ColumnDefinitions>
-                    <Button Grid.Column="0" Command="{Binding MoveUpCategoryItemCommand}" CommandParameter="{Binding SelectedIndex, ElementName=CategoryListBox}" Margin="0,2.5,0,0">
+                    <Button Grid.Column="0" Command="{Binding MoveUpCategoryItemCommand}" CommandParameter="{Binding SelectedIndex, ElementName=CategoryListBox}" Margin="0,0,0,2.5">
                         <Grid>
                             <Grid.ColumnDefinitions>
                                 <ColumnDefinition Width="*"></ColumnDefinition>
@@ -80,7 +80,7 @@
                             <TextBlock Grid.Column="1" Text="Subir" VerticalAlignment="Center" Foreground="Gray" MinWidth="100"></TextBlock>
                         </Grid>
                     </Button>
-                    <Button Grid.Column="1" Command="{Binding MoveDownCategoryItemCommand}" CommandParameter="{Binding SelectedIndex, ElementName=CategoryListBox}" Margin="0,2.5,0,0">
+                    <Button Grid.Column="1" Command="{Binding MoveDownCategoryItemCommand}" CommandParameter="{Binding SelectedIndex, ElementName=CategoryListBox}" Margin="0,0,0,2.5">
                         <Grid>
                             <Grid.ColumnDefinitions>
                                 <ColumnDefinition Width="*"></ColumnDefinition>
