+++ View/SystemPage.xaml
@@ -113,7 +113,14 @@
                 <ColumnDefinition Width="*"></ColumnDefinition>
             </Grid.ColumnDefinitions>
             <!-- System version -->
-            <TextBlock Foreground="LightGray" FontSize="18" HorizontalAlignment="Center" Text="{Binding Version}"></TextBlock>
+            <Grid Grid.Row="0">
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition Width="*"></ColumnDefinition>
+                    <ColumnDefinition Width="*"></ColumnDefinition>
+                </Grid.ColumnDefinitions>
+                <TextBlock Grid.Column="0" Foreground="LightGray" FontSize="18" HorizontalAlignment="Right" Text="Wibsar Retail V."></TextBlock>
+                <TextBlock Grid.Column="1" Foreground="LightGray" FontSize="18" HorizontalAlignment="Left" Margin="5,0,0,0" Text="{Binding Version}"></TextBlock>
+                </Grid>
             <!-- Save Button -->
             <Button Grid.Column="2" Command="{Binding SystemSaveChangesCommand}"  Background="{x:Null}" BorderBrush="{x:Null}">
                 <Grid>
